Updated DTOs and Business logic

+++ BACKEND/girls_for_girls/src/main/java/com/neobis/g4g/girls_for_girls/data/dto/TokenDTO.java
@@ -6,7 +6,7 @@ import lombok.Setter;
 @Getter
 @Setter
 public class TokenDTO {
-    private Integer userId;
+    private long userId;
     private String accessToken;
     private String refreshToken;
 }
